Move BaseCommon back to tests file for better visibility
